Add files via upload
The future scope of the QuickFarm ecommerce website increased use of e-commerce and 
organic farming businesses the QuickFarm prospects are very bright. However, overall, after 
the economic market collapse, the potential for future boom and development in e-commerce 
has definite possibilities, including the QuickFarm e-commerce platform. The scenario may be 
also explained due to the growing importance of the role of e-commerce in the modern society 
that enables QuickFarm to evolve and implement better features and functions that are more 
responsive to the needs of the customers. This implies that users would have more than one 
billing and shipping address and they would simply be able to select the relevant billing and 
shipping addresses during the checkout. Users could subscribe to price alerts that would allow 
them to receive notifications when the price of particular products fell to a certain level. Expand 
the product range to include, in addition to organic and specialized products, also farming and 
gardening tools and supplies.